+++ code/+zapit/+gui/+main/@controller/saveGUIcache.m
@@ -1,0 +1,9 @@
+function saveGUIcache(obj)
+    % Save the GUI cache
+
+    zapitGUIcache.previouslyLoadedStimConfigs = obj.previouslyLoadedStimConfigs;
+
+    fname = obj.GUIcacheLocation;
+    save(fname,'zapitGUIcache')
+
+end