+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The project will adhere to [semantic versioning](http://semver.org) guidelines, 
 2023/01/XX -- VERSION
  * Move relevant methods into zapit.stimConfig
  * zapit.stimConfig.makeChanSamples is now turned into a getter of chanSamples
+ * Recent files updates when files are missing. The list is cached and re-appears on reload.
 
 
 2023/01/12 -- v0.5.1-alpha